Fixed serious AI bug where a major nation was not mutually in war with an indy civ in some special cases Fixed bug where later set (on other civs spawn for example) potential provinces were never upgraded to historical/core Fixed bug where all provinces of the given civ were reset to their initial status on respawn (thus not taking later changes into account) Fixed issue on respawn where potential provinces didn't change status with the flipped cities Found provincetype issue connected to province initialization, added some asserts for it until provinces are updated Fixed assert failure connected to the new civic screen Fixed possible CTD with some AI worker action calculations, based on a Firaxis mistake - by OrionVeteran Fixed another (probably harmless) assert failure, connected to civ initialization Fixed bug in unit looping connected to prosecutors, which results in a huge performance improvement - thanks to Nightinggale Fixed bug where Byzantium could collapse in some rare cases even without losing Constantinople Fixed various bugs in the Venetian preset war against independent Ragusa Fixed bug (and potential crash) connected to renaming the city on city secession Fixed bug where religion foundation date were displayed as "Early Middle Ages" for almost all game long Fixed bug: the Pope now always ignores Barbs and Indies outside it's war map
Minor text updates (DCN, civilopedia, message, hint)
Fixed some mistakes in the French and Venetian settler maps
Switched colours of spawn and respawn area displays
Religion Screen and Company Screen can now handle unicode characters
Improved religion foundation date information on the Religion Screen, also it now takes different scenarios into account
Improved city and wonder foundation date/city/civ information on the Info Screen, also it now takes different scenarios into account
Fixed mistakes on the Info Screen where information from the top5 cities and from the wonders were not in sync (connected to city visibility and contact with the civ)
Readded era displays to various screens (main, info, religion) before the discovery of Calendar

Respawned civs can no longer collapse in the first 10 turns
No city secession in the first 15 turns after spawn, or in the first 10 turns after respawn
If the civ is respawned, it's harder to collapse them by barbs
Vassals can now also instantly collapse if they lose a significant amount of their territory in a couple turns
Fixed issue where the human player collapsed less often then it should
Improved mechanics for all 4 different collapse types, collapses should be way more straightforward
Increased frequency of resurrections

Correct CNM name for new cities placed in the WB
Since the WB file cannot handle special characters and long names perfectly, added function to rename city names on the start of the scenario
Reduced overall damage from the Hungarian Végvár: units won't get a fixed amount of damage, it depends on their current HP instead (so Végvár won't kill units, only weaken them)
AI civs will more often reconsider their wars against minor civs (indies)
AI war declaration on indies is now based on their war maps, the chance is increased with higher war map values
AI civs will much more frequently declare war on indies in their core area
AI declares war on the new indy city right after city secession
Only half of the units present in the city will flip to the new indy civ/city on secession

+++ RFCEurope/Assets/Python/RFCUtils.py
@@ -2,7 +2,7 @@
 
 from CvPythonExtensions import *
 import CvUtil
-import CvScreenEnums #Absinhe
+import CvScreenEnums #Absinthe
 import RFCEMaps #Absinthe
 import PyHelpers
 import Popup #Absinthe
@@ -47,6 +47,12 @@ class RFCUtils:
 
 	def setLastTurnAlive( self, iCiv, iNewValue ):
 		sd.scriptDict['lLastTurnAlive'][iCiv] = iNewValue
+
+	def getLastRespawnTurn( self, iCiv ):
+		return sd.scriptDict['lLastRespawnTurn'][iCiv]
+
+	def setLastRespawnTurn( self, iCiv, iNewValue ):
+		sd.scriptDict['lLastRespawnTurn'][iCiv] = iNewValue
 
 	#Victory
 	#def getGoal( self, i, j ):
@@ -313,7 +319,7 @@ class RFCUtils:
 							teamMinor.signOpenBorders(iActiveCiv)
 
 	#AIWars
-	def restorePeaceHuman(self, iMinorCiv, bOpenBorders):
+	def restorePeaceHuman(self, iMinorCiv):
 		teamMinor = gc.getTeam(gc.getPlayer(iMinorCiv).getTeam())
 		for iActiveCiv in range( iNumActivePlayers ):
 			if (gc.getPlayer(iActiveCiv).isHuman()):
@@ -326,7 +332,7 @@ class RFCUtils:
 				return
 
 	#AIWars
-	def minorWars(self, iMinorCiv):
+	def minorWars(self, iMinorCiv, iGameTurn):
 		teamMinor = gc.getTeam(gc.getPlayer(iMinorCiv).getTeam())
 		apCityList = PyPlayer(iMinorCiv).getCityList()
 		for pCity in apCityList:
@@ -334,13 +340,44 @@ class RFCUtils:
 			x = city.getX()
 			y = city.getY()
 			for iActiveCiv in range( iNumActivePlayers ):
-				# 3MiroPapal: the Pope does not attack independent
 				if ( gc.getPlayer(iActiveCiv).isAlive() and (not gc.getPlayer(iActiveCiv).isHuman()) and (not iActiveCiv == con.iPope) ):
-					if (gc.getPlayer(iActiveCiv).getSettlersMaps( con.iMapMaxY-y-1, x ) >= 90):
-						if (not teamMinor.isAtWar(iActiveCiv)):
-							teamMinor.declareWar(iActiveCiv, False, WarPlanTypes.WARPLAN_LIMITED)
-							print ("Minor war", city.getName(), gc.getPlayer(iActiveCiv).getCivilizationAdjective(0))
+					if (not teamMinor.isAtWar(iActiveCiv)):
+						if (iGameTurn > con.tBirth[iActiveCiv] + 20):
+							# Absinthe: probably better to use war maps instead of settler maps, but let the AI concentrate on it's core area first
+							#			maybe we should use both settler and war maps? distance calculations would be great, but use too much iterations
+							#if (gc.getPlayer(iActiveCiv).getSettlersMaps( con.iMapMaxY-y-1, x ) >= 90 or gc.getPlayer(iActiveCiv).getSettlersMaps( con.iMapMaxY-y-1, x ) == -1):
+							#if (gc.getPlayer(iActiveCiv).getWarsMaps( con.iMapMaxY-y-1, x ) >= 2):
+							iRndNum = gc.getGame().getSorenRandNum( 10, 'random warmap chance')
+							if (gc.getPlayer(iActiveCiv).getWarsMaps( con.iMapMaxY-y-1, x ) >= 10):
+								# 100% chance for cities with high war map value
+								teamMinor.declareWar(iActiveCiv, False, WarPlanTypes.WARPLAN_LIMITED)
+								print ("minorWars", city.getName(), gc.getPlayer(iActiveCiv).getCivilizationAdjective(0), gc.getPlayer(iActiveCiv).getSettlersMaps( con.iMapMaxY-y-1, x ), gc.getPlayer(iActiveCiv).getWarsMaps( con.iMapMaxY-y-1, x ), "WARPLAN_LIMITED")
+							elif (gc.getPlayer(iActiveCiv).getWarsMaps( con.iMapMaxY-y-1, x ) >= 6):
+								if (iRndNum < 8): # 80% chance for cities with medium war map value
+									teamMinor.declareWar(iActiveCiv, False, WarPlanTypes.WARPLAN_LIMITED)
+									print ("minorWars", city.getName(), gc.getPlayer(iActiveCiv).getCivilizationAdjective(0), gc.getPlayer(iActiveCiv).getSettlersMaps( con.iMapMaxY-y-1, x ), gc.getPlayer(iActiveCiv).getWarsMaps( con.iMapMaxY-y-1, x ), "WARPLAN_LIMITED")
+							elif (gc.getPlayer(iActiveCiv).getWarsMaps( con.iMapMaxY-y-1, x ) >= 2):
+								if (iRndNum < 5): # 50% chance for cities with low war map value
+									teamMinor.declareWar(iActiveCiv, False, WarPlanTypes.WARPLAN_LIMITED)
+									print ("minorWars", city.getName(), gc.getPlayer(iActiveCiv).getCivilizationAdjective(0), gc.getPlayer(iActiveCiv).getSettlersMaps( con.iMapMaxY-y-1, x ), gc.getPlayer(iActiveCiv).getWarsMaps( con.iMapMaxY-y-1, x ), "WARPLAN_LIMITED")
 
+	#AIWars
+	# Absinthe: declare war sooner / more frequently if an Indy city has huge value on the civ's war map
+	def minorCoreWars(self, iMinorCiv, iGameTurn):
+		teamMinor = gc.getTeam(gc.getPlayer(iMinorCiv).getTeam())
+		apCityList = PyPlayer(iMinorCiv).getCityList()
+		for pCity in apCityList:
+			city = pCity.GetCy()
+			x = city.getX()
+			y = city.getY()
+			for iActiveCiv in range( iNumActivePlayers ):
+				if ( gc.getPlayer(iActiveCiv).isAlive() and (not gc.getPlayer(iActiveCiv).isHuman()) and (not iActiveCiv == con.iPope) ):
+					# Absinthe: do not want to force the AI into these wars with WARPLAN_TOTAL too early
+					if (iGameTurn > con.tBirth[iActiveCiv] + 40):
+						if (not teamMinor.isAtWar(iActiveCiv)):
+							if (gc.getPlayer(iActiveCiv).getWarsMaps( con.iMapMaxY-y-1, x ) == 16):
+								teamMinor.declareWar(iActiveCiv, False, WarPlanTypes.WARPLAN_TOTAL)
+								print ("minorWars", city.getName(), gc.getPlayer(iActiveCiv).getCivilizationAdjective(0), gc.getPlayer(iActiveCiv).getSettlersMaps( con.iMapMaxY-y-1, x ), gc.getPlayer(iActiveCiv).getWarsMaps( con.iMapMaxY-y-1, x ), "WARPLAN_TOTAL")
 
 	#RiseAndFall, Stability
 	def calculateDistance(self, x1, y1, x2, y2):
@@ -390,9 +427,15 @@ class RFCUtils:
 			unit = plotCity.getUnit(j)
 			unitType = unit.getUnitType()
 			if (unit.getOwner() == iOldOwner):
-				unit.kill(False, con.iBarbarian)
-				if (iNewOwner < con.iNumActivePlayers or unitType > xml.iSettler):
-					self.makeUnit(unitType, iNewOwner, [0, 72], 1)
+				# Absinthe: only 50% chance that the unit will defect
+				if ( gc.getGame().getSorenRandNum(2, 'Convert Unit') == 1 ):
+					unit.kill(False, con.iBarbarian)
+					if (iNewOwner < con.iNumActivePlayers or unitType > xml.iSettler):
+						self.makeUnit(unitType, iNewOwner, [0, 72], 1)
+				# skip unit if it won't defect, so it will move out of the city territory
+				else:
+					j += 1
+			# skip unit if from another player
 			else:
 				j += 1
